分页 + 搜索 例子

+++ src/main/java/net/masadora/mall/framework/interfaces/rest/LeylineDomainRestCRUD.java
@@ -28,7 +28,7 @@ public abstract class LeylineDomainRestCRUD<T extends LeylineRepo, D extends Ley
 
     @SuppressWarnings(value = "unchecked")
     @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = "application/json")
-    public D find(Integer id) {
+    public D find(Long id) {
         return (D) dao.findOne(id);
     }
